+++ packages/evokit-picture/CHANGELOG.md
@@ -6,9 +6,19 @@
 
 ### CHANGELOG
 
+#### 3.0.2 *(11.07.2019)*
+
+- **Added** new advanced border props:
+    - `picture-border-top`
+    - `picture-border-right`
+    - `picture-border-bottom`
+    - `picture-border-left`
+    - `picture-border-tb`
+    - `picture-border-lr`
+
 #### 3.0.1 *(19.06.2019)*
 
-- Added export `PictureItem`, for example `import { Picture, PictureItem } from 'evokit-picture';`
+- **Added** export `PictureItem`, for example `import { Picture, PictureItem } from 'evokit-picture';`
 
 #### 3.0.0